backupStudy {antaresEditObject}	R Documentation

Create a backup with an Antares Study

Description

Save an Antares Study or only inputs in a .tar.gz file

Usage

backupStudy(
  backupfile,
  what = c("input", "study"),
  opts = antaresRead::simOptions()
)

Arguments

backupfile
Name of the backup, without extension. If missing, either the name of the study or 'input' according argument what.

what
Which folder to save, input for the input folder or study for the whole study.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

The path of the backup

Examples

## Not run: 
backupStudy()

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]