checkRemovedArea {antaresEditObject}	R Documentation

Seek for a removed area

Description

Check if it remains trace of a deleted area in the input folder

Usage

checkRemovedArea(area, all_files = TRUE, opts = antaresRead::simOptions())

Arguments

area
An area

all_files
Check files in study directory.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

a named list with two elements

Examples

## Not run: 
checkRemovedArea("myarea")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]