copyOutput {antaresEditObject}	R Documentation

Copy of the output files of an Antares study

Description

Copy of the output files of an Antares study

Usage

copyOutput(opts, extname, mcYears = "all")

Arguments

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

extname
Extension to be added to the name of the study, to be used as a name for the newly created folder.

mcYears
mcYears to copy. Can be "all".

Examples

## Not run: 

library(antaresRead)

# Set simulation path
opts = setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Create a new area
copyOutput(opts, "_adq")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]