createArea {antaresEditObject}	R Documentation

Create An Area In An Antares Study

Description

Create An Area In An Antares Study

Usage

createArea(
  name,
  color = grDevices::rgb(230, 108, 44, max = 255),
  localization = c(0, 0),
  nodalOptimization = nodalOptimizationOptions(),
  filtering = filteringOptions(),
  overwrite = FALSE,
  opts = antaresRead::simOptions()
)

Arguments

name
Name of the area as a character, without punctuation except - and _.

color
Color of the node

localization
Localization on the map

nodalOptimization
Nodal optimization parameters, see nodalOptimizationOptions.

filtering
Filtering parameters, see filteringOptions.

overwrite
Overwrite the area if already exist.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

## Not run: 

library(antaresRead)

# Set simulation path
setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Create a new area
createArea("fictive_area")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]