Create a link between two areas

Create a link between two areas

createLink(
  from,
  to,
  propertiesLink = propertiesLinkOptions(),
  dataLink = NULL,
  overwrite = FALSE,
  opts = antaresRead::simOptions()
)

Arguments

from
The first area from which to create a link

to
The second one

propertiesLink
a named list containing the link properties, e.g. hurdles-cost or transmission-capacities for example. See propertiesLinkOptions.

dataLink
For Antares v7, a matrix with eight column corresponding to : trans. capacity (direct) trans. capacity (indirect), hurdles cost (direct), hurdles cost (indirect), impedances, loop flow, PST min, PST max. If NULL (default), a matrix whose rows are equal to 1, 1, 0, 0, 0, 0, 0, 0 is set. See Details

overwrite
Logical, overwrite the previous between the two areas if exist

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Details

The eight times-series are:

NTC direct : the upstream-to-downstream capacity, in MW

NTC indirect : the downstream-to-upstream capacity, in MW

Hurdle cost direct : an upstream-to-downstream transmission fee, in euro/MWh

Hurdle cost indirect : a downstream-to-upstream transmission fee, in euro/MWh

Impedances : virtual impedances that are used in economy simulations to give a physical meaning to raw outputs, when no binding constraints have been defined to enforce Kirchhoff's laws.

Loop flow : amount of power flowing circularly though the grid when all "nodes" are perfectly balanced (no import and no export).

PST min : lower bound of phase-shifting that can be reached by a PST installed on the link, if any.

PST max : upper bound of phase-shifting that can be reached by a PST installed on the link, if any.

NB: For Antares v7 the eight columns must conform to above order. For Antares v6, only five columns are expected, and they must follow this other order: NTC direct, NTC indirect, Impedances, Hurdle cost direct, Hurdle cost indirect.

Note

In Antares, areas are sorted in alphabetical order to establish links between. For example, link between "fr" and "be" will appear under "be". So the areas are sorted before creating the link between them, and dataLink is rearranged to match the new order.

Examples

if (FALSE) {

library(antaresRead)

# Set simulation path
setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Create a link between two areas
createLink(from = "first_area", to  = "second_area")

}