edit-cluster {antaresEditObject}	R Documentation

Edit an existing cluster

Description

Edit parameters of an existing cluster, thermal or RES (renewable energy source).

Usage

editCluster(
  area,
  cluster_name,
  ...,
  time_series = NULL,
  prepro_data = NULL,
  prepro_modulation = NULL,
  add_prefix = TRUE,
  opts = antaresRead::simOptions()
)

editClusterRES(
  area,
  cluster_name,
  ...,
  time_series = NULL,
  add_prefix = TRUE,
  opts = antaresRead::simOptions()
)

Arguments

area
The area where the cluster is.

cluster_name
cluster name.

...
Parameters to write in the Ini file.

time_series
the "ready-made" 8760-hour time-series available for simulation purposes.

prepro_data
Pre-process data, a data.frame or matrix, default is a matrix with 365 rows and 6 columns.

prepro_modulation
Pre-process modulation, a data.frame or matrix, if specified, must have 8760 rows and 1 or 4 columns.

add_prefix
If TRUE, cluster_name will be prefixed by area name.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath()

Value

An updated list containing various information about the simulation.

See Also

createCluster() or createClusterRES() to create new clusters, removeCluster() or removeClusterRES() to remove clusters.

Examples

## Not run: 

# Update only nominalCapacity for an existing cluster
editCluster(
  area = "myarea", 
  cluster_name = "mycluster", 
  nominalcapacity = 10600.000
)

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]