editBindingConstraint {antaresEditObject}	R Documentation

Update An Existing Binding Constraint

Description

Update An Existing Binding Constraint

Usage

editBindingConstraint(
  name,
  id = tolower(name),
  values = NULL,
  enabled = NULL,
  timeStep = NULL,
  operator = NULL,
  coefficients = NULL,
  opts = antaresRead::simOptions()
)

Arguments

name
The name for the binding constraint

id
An id

values
Values used by the constraint. It contains one line per time step and three columns "less", "greater" and "equal".

enabled
Logical, is the constraint enabled ?

timeStep
Time step the constraint applies to : hourly, daily or weekly

operator
Type of constraint: equality, inequality on one side or both sides.

coefficients
A named vector containing the coefficients used by the constraint.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

## Not run: 
editBindingConstraint(
  name = "toto", 
  values = matrix(data = rep(0, 8760 * 3), ncol = 3), 
  enabled = FALSE, 
  timeStep = "hourly",
  operator = "both",
  coefficients = c("fr%de" = 1)
)

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]