editLink {antaresEditObject}	R Documentation

Edit a link between two areas

Description

Edit a link between two areas

Usage

editLink(
  from,
  to,
  hurdles_cost = NULL,
  transmission_capacities = NULL,
  asset_type = NULL,
  display_comments = NULL,
  filter_synthesis = NULL,
  filter_year_by_year = NULL,
  dataLink = NULL,
  opts = antaresRead::simOptions()
)

Arguments

from
The first area from which to create a link

to
The second one

hurdles_cost
Logical, which is used to state whether (linear) transmission fees should be taken into account or not in economy and adequacy simulations

transmission_capacities
Character, one of enabled, ignore or infinite, which is used to state whether the capacities to consider are those indicated in 8760-hour arrays or if zero or infinite values should be used instead (actual values / set to zero / set to infinite)

asset_type
Character, one of ac, dc, gas, virt or other. Used to state whether the link is either an AC component (subject to Kirchhoff’s laws), a DC component, or another type of asset.

display_comments
Logical

filter_synthesis
Output synthesis

filter_year_by_year
Output year-by-year

dataLink
For Antares v7, a matrix with eight column corresponding to : trans. capacity (direct) trans. capacity (indirect), hurdles cost (direct), hurdles cost (indirect), impedances, loop flow, PST min, PST max. If NULL (default), a matrix whose rows are equal to 1, 1, 0, 0, 0, 0, 0, 0 is set. See Details

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Details

The eight times-series are:

NTC direct : the upstream-to-downstream capacity, in MW

NTC indirect : the downstream-to-upstream capacity, in MW

Hurdle cost direct : an upstream-to-downstream transmission fee, in euro/MWh

Hurdle cost indirect : a downstream-to-upstream transmission fee, in euro/MWh

Impedances : virtual impedances that are used in economy simulations to give a physical meaning to raw outputs, when no binding constraints have been defined to enforce Kirchhoff's laws.

Loop flow : amount of power flowing circularly though the grid when all "nodes" are perfectly balanced (no import and no export).

PST min : lower bound of phase-shifting that can be reached by a PST installed on the link, if any.

PST max : upper bound of phase-shifting that can be reached by a PST installed on the link, if any.

NB: For Antares v7 the eight columns must conform to above order. For Antares v6, only five columns are expected, and they must follow this other order: NTC direct, NTC indirect, Impedances, Hurdle cost direct, Hurdle cost indirect.

Value

An updated list containing various information about the simulation.

Note

In Antares, areas are sorted in alphabetical order to establish links between. For example, link between "fr" and "be" will appear under "be". So the areas are sorted before creating the link between them, and dataLink is rearranged to match the new order.

Examples

## Not run: 
editLink(
  from = "area1",
  to = "area2",
  transmission_capacities = "infinite"
)

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]