getPlaylist {antaresEditObject}	R Documentation

Get the playlist of an Antares study

Description

getPlaylist gives the identifier of the MC years which will be simulated in the Antares study, taking into account the potential use of a playlist which can skip some MC years

Usage

getPlaylist(opts = antaresRead::simOptions())

Arguments

opts
list of simulation parameters returned by the function antaresRead::setSimulationPath

Value

Returns a vector of the identifier of the simulated MC year

[Package antaresEditObject version 0.3.0 Index]