remove-cluster {antaresEditObject}	R Documentation

Remove a cluster

Description

Remove a cluster, thermal or RES (renewable energy source), and all its data.

Usage

removeCluster(
  area,
  cluster_name,
  add_prefix = TRUE,
  opts = antaresRead::simOptions()
)

removeClusterRES(
  area,
  cluster_name,
  add_prefix = TRUE,
  opts = antaresRead::simOptions()
)

Arguments

area
Area from which to remove a cluster.

cluster_name
Cluster to remove.

add_prefix
If TRUE, cluster_name will be prefixed by area's name.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

See Also

createCluster() or createClusterRES() to create new clusters, editCluster() or editClusterRES() to edit existing clusters.

Examples

## Not run: 
createCluster(area = "fr", cluster_name = "fr_gas",
              group = "other", `marginal-cost` = 50)

removeCluster(area = "fr", cluster_name = "fr_gas")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]