removeArea {antaresEditObject}	R Documentation

Remove An Area From inputs

Description

Remove An Area From inputs

Usage

removeArea(name, opts = antaresRead::simOptions())

Arguments

name
An area name

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

## Not run: 
removeArea("fictive_area")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]