removeBindingConstraint {antaresEditObject}	R Documentation

Remove a Binding Constraint

Description

Remove a Binding Constraint

Usage

removeBindingConstraint(name, opts = antaresRead::simOptions())

Arguments

name
Name(s) of the binding constraint(s) to remove.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

## Not run: 
removeBindingConstraint("mybindingconstraint")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]