removeLink {antaresEditObject}	R Documentation

Remove a link between two areas

Description

Remove a link between two areas

Usage

removeLink(from, to, opts = antaresRead::simOptions())

Arguments

from
The first area from which to create a link

to
The second one

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

## Not run: 
createLink(from = "myarea", to  = "myarea2")
removeLink(from = "myarea", to  = "myarea2")

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]