runSimulation {antaresEditObject}	R Documentation

Run an Antares Simulation

Description

run_simulation is a function which runs an ANTARES study in economic mode

Usage

runSimulation(
  name,
  mode = "economy",
  path_solver = getOption("antares.solver"),
  wait = TRUE,
  show_output_on_console = FALSE,
  parallel = TRUE,
  opts = antaresRead::simOptions()
)

Arguments

name
Name of the simulation.

mode
Simulation mode, can take value "economy", "adequacy" or "draft".

path_solver
Character containing the Antares Solver path

wait
Logical, indicating whether the R interpreter should wait for the simulation to finish, or run it asynchronously.

show_output_on_console
Logical, indicating whether to capture the ANTARES log and show it on the R console.

parallel
Logical. If TRUE the ANTARES simulation will be run in parallel mode (Work only with ANTARES v6.0.0 or more). In that case, the number of cores used by the simulation is the one set in advanced_settings/simulation_cores (see ANTARES interface).

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

The function does not return anything. It is used to launch an ANTARES simulation

[Package antaresEditObject version 0.3.0 Index]