runTsGenerator {antaresEditObject}	R Documentation

Run Time-Series Generator

Description

Run Time-Series Generator

Usage

runTsGenerator(
  path_solver = getOption("antares.solver"),
  wait = TRUE,
  show_output_on_console = FALSE,
  opts = antaresRead::simOptions()
)

Arguments

path_solver
Character containing the Antares Solver path.

wait
Logical, indicating whether the R interpreter should wait for the simulation to finish, or run it asynchronously.

show_output_on_console
Logical, indicating whether to capture the ANTARES log and show it on the R console.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath.

Examples

## Not run: 
library(antaresRead)
setSimulationPath(path = "path/to/study")

library(antaresEditObject)
runTsGenerator(
  path_solver = "path/to/antares-6.0-solver.exe", 
  show_output_on_console = TRUE
)

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]