setPlaylist {antaresEditObject}	R Documentation

Set the playlist of an Antares Study

Description

set_playlist is a function which modifies the input file of an ANTARES study and set the playlist in order to simulate only the MC years given in input

Usage

setPlaylist(playlist, weights = NULL, opts = antaresRead::simOptions())

Arguments

playlist
vector of MC years identifier to be simulated can be a list (V8 compatibility) but not recommended

weights
data.table, 2 columns : mcYears and weights. Only with after antares V8

opts
list of simulation parameters returned by the function antaresRead::setSimulationPath

Value

The function does not return anything. It is used to modify the input of an Antares study

[Package antaresEditObject version 0.3.0 Index]