updateInputSettings {antaresEditObject}	R Documentation

Update input parameters of an Antares study

Description

Update input parameters of an Antares study

Usage

updateInputSettings(import, opts = antaresRead::simOptions())

Arguments

import
Series to import.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath()

Value

An updated list containing various information about the simulation.

Examples

## Not run: 

updateInputSettings(import = c("thermal"))
updateInputSettings(import = c("hydro", "thermal"))

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]