updateOptimizationSettings {antaresEditObject}	R Documentation

Update optimization parameters of an Antares study

Description

Update optimization parameters of an Antares study

Usage

updateOptimizationSettings(
  simplex.range = NULL,
  transmission.capacities = NULL,
  include.constraints = NULL,
  include.hurdlecosts = NULL,
  include.tc.min.stable.power = NULL,
  include.tc.min.up.down.time = NULL,
  include.dayahead = NULL,
  include.strategicreserve = NULL,
  include.spinningreserve = NULL,
  include.primaryreserve = NULL,
  include.exportmps = NULL,
  power.fluctuations = NULL,
  shedding.strategy = NULL,
  shedding.policy = NULL,
  unit.commitment.mode = NULL,
  number.of.cores.mode = NULL,
  renewable.generation.modelling = NULL,
  day.ahead.reserve.management = NULL,
  opts = antaresRead::simOptions()
)

Arguments

simplex.range
week or day

transmission.capacities
true, false or infinite

include.constraints
true or false

include.hurdlecosts
true or false

include.tc.min.stable.power
true or false

include.tc.min.up.down.time
true or false

include.dayahead
true or false

include.strategicreserve
true or false

include.spinningreserve
true or false

include.primaryreserve
true or false

include.exportmps
true or false

power.fluctuations
free modulations, minimize excursions or minimize ramping

shedding.strategy
share margins

shedding.policy
shave peaks or minimize duration

unit.commitment.mode
fast or accurate

number.of.cores.mode
minimum, low, medium, high or maximum

renewable.generation.modelling
aggregated or clusters

day.ahead.reserve.management
global

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation options.

[Package antaresEditObject version 0.3.0 Index]