updateOutputSettings {antaresEditObject}	R Documentation

Update output parameters of an Antares study

Description

Update output parameters of an Antares study

Usage

updateOutputSettings(
  synthesis = NULL,
  storenewset = NULL,
  archives = NULL,
  opts = antaresRead::simOptions()
)

Arguments

synthesis
Logical. If TRUE, synthetic results will be stored in a directory Study_name/OUTPUT/simu_tag/Economy/mc-all. If FALSE, No general synthesis will be printed out.

storenewset
Logical. See Antares General Reference Guide.

archives
Character vector. Series to archive.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

## Not run: 

updateOutputSettings(synthesis = TRUE, storenewset = FALSE,
                     archives = c("load", "wind"))

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]