writeInputTS {antaresEditObject}	R Documentation

Write input time series

Description

This function writes input time series in an Antares project.

Usage

writeInputTS(
  area,
  type = c("load", "hydroROR", "hydroSTOR", "wind", "solar"),
  data,
  overwrite = TRUE,
  opts = antaresRead::simOptions()
)

Arguments

area
The area where to write the input time series.

type
Serie to write: "load", "hydroROR", "hydroSTOR", "wind" or "solar".

data
A 8760N matrix of hourly time series, except when type is "hydroSTOR". In this latter case, data must either be 365N (Antares v7) or 12*N (v6 and earlier).

overwrite
Logical. Overwrite the values if a file already exists.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath.

Examples

## Not run: 

writeInputTS("fictive_area", type = "solar", data = matrix(rep(4, 8760*2), nrow = 8760))

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]