writeOutputValues {antaresEditObject}	R Documentation

Write output value for Antares

Description

This function write all output values for an Antares study.

Usage

writeOutputValues(data, opts = NULL)

Arguments

data
obtain with readAntares

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Examples

## Not run: 

library(antaresRead)
library(data.table)
opts <- setSimulationPath("my_study")
data <- readAntares(links = "all", areas = "all", clusters = "all")
writeOutputValues(data)

## End(Not run)

[Package antaresEditObject version 0.3.0 Index]