Create a cluster

Create a new thermal or RES (renewable energy source) cluster.

createCluster(
  area,
  cluster_name,
  group = "Other",
  ...,
  time_series = NULL,
  prepro_data = NULL,
  prepro_modulation = NULL,
  add_prefix = TRUE,
  overwrite = FALSE,
  opts = antaresRead::simOptions()
)

createClusterRES(
  area,
  cluster_name,
  group = "Other RES 1",
  ...,
  time_series = NULL,
  add_prefix = TRUE,
  overwrite = FALSE,
  opts = antaresRead::simOptions()
)

Arguments

area
The area where to create the cluster.

cluster_name
Name for the cluster, it will prefixed by area name, unless you set add_prefix = FALSE.

group
Group of the cluster, depends on cluster type:

thermal cluster, one of: Gas, Hard coal, Lignite, Mixed fuel, Nuclear, Oil, Other, Other 2, Other 3, Other 4.

renewable cluster, one of: Wind Onshore, Wind Offshore, Solar Thermal, Solar PV, Solar Rooftop, Other RES 1, Other RES 2, Other RES 3, Other RES 4.

...
Parameters to write in the Ini file. Careful! Some parameters must be set as integers to avoid warnings in Antares, for example, to set unitcount, you'll have to use unitcount = 1L.

time_series
the "ready-made" 8760-hour time-series available for simulation purposes.

prepro_data
Pre-process data, a data.frame or matrix, default is a matrix with 365 rows and 6 columns.

prepro_modulation
Pre-process modulation, a data.frame or matrix, if specified, must have 8760 rows and 1 or 4 columns.

add_prefix
If TRUE (the default), cluster_name will be prefixed by area name.

overwrite
Logical, overwrite the cluster or not.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath()

Value

An updated list containing various information about the simulation.

See also

Examples