Edit An Existing Area In An Antares Study

Edit An Existing Area In An Antares Study

editArea(
  name,
  color = NULL,
  localization = NULL,
  nodalOptimization = NULL,
  filtering = NULL,
  opts = antaresRead::simOptions()
)

Arguments

name
Name of the area as a character, without punctuation except - and _.

color
Color of the node

localization
Localization on the map

nodalOptimization
Nodal optimization parameters, see nodalOptimizationOptions.

filtering
Filtering parameters, see filteringOptions.

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

Value

An updated list containing various information about the simulation.

Examples

if (FALSE) {

library(antaresRead)

# Set simulation path
setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Edit an existing area
editArea("area", color = grDevices::rgb(230, 108, 44, max = 255),
  localization = c(1, 1),
  opts = antaresRead::simOptions()) 

editArea("de",  nodalOptimization = list("spilledenergycost" = list(fr = 30)),
opts = antaresRead::simOptions())

editArea("de",  nodalOptimization = nodalOptimizationOptions(),
opts = antaresRead::simOptions())

}