Read, create & update scenario builder

Read, create & update scenario builder

scenarioBuilder(
  n_scenario,
  n_mc = NULL,
  areas = NULL,
  areas_rand = NULL,
  opts = antaresRead::simOptions()
)

readScenarioBuilder(
  ruleset = "Default Ruleset",
  as_matrix = TRUE,
  opts = antaresRead::simOptions()
)

updateScenarioBuilder(
  ldata,
  ruleset = "Default Ruleset",
  series = NULL,
  opts = antaresRead::simOptions()
)

Arguments

n_scenario
Number of scenario.

n_mc
Number of Monte-Carlo years.

areas
Areas to use in scenario builder, if NULL (default), areas in Antares study are used.

areas_rand
Areas for which to use "rand".

opts
List of simulation parameters returned by the function antaresRead::setSimulationPath

ruleset
Ruleset to read.

as_matrix
If TRUE (default) return a matrix, else a list.

ldata
A matrix obtained with scenarioBuilder, or a named list of matrices obtained with scenarioBuilder, names must be 'l', 'h', 'w', 's' or 't', depending on the series to update.

series
Name(s) of the serie(s) to update if ldata is a single matrix.

Value

scenarioBuilder : a matrix

readScenarioBuilder : a list of matrix or list

Examples

if (FALSE) {

library(antaresRead)
library(antaresEditObject)

# simulation path
setSimulationPath(
  path = "pat/to/simulation",
  simulation = "input"
)

# Create a scenario builder matrix
sbuilder <- scenarioBuilder(
  n_scenario = 51,
  n_mc = 2040,
  areas_rand = c("fr", "be")
)
sbuilder[, 1:6]
dim(sbuilder)

# Read previous scenario builder
# in a matrix format
prev_sb <- readScenarioBuilder()

# Update scenario builder

# for load serie
updateScenarioBuilder(ldata = sbuilder, series = "load")

# equivalent as
updateScenarioBuilder(ldata = list(l = sbuilder))

# update several series

# same input
sbuilder
updateScenarioBuilder(
  ldata = sbuilder,
  series = c("load", "hydro", "solar")
)

# different input
updateScenarioBuilder(ldata = list(
  l = load_sb,
  h = hydro_sb,
  s = solar_sb
))

}